Comment functions

+++ back/gupi_api/src/main/java/com/skywon/gupi/controller/UserController.java
@@ -23,7 +23,6 @@ public class UserController {
 
     /**
      * Récupérer tous les utilisateurs
-     * C
      * @return
      */
     @GetMapping("/")
@@ -31,11 +30,22 @@ public class UserController {
         return this.userService.getAll();
     }
 
+    /**
+     * Récupérer un seul utilisateur
+     * @param id
+     * @return
+     */
     @GetMapping("/{id}")
     public User getUser(@PathVariable Integer id) {
         return this.userService.byId(id);
     }
 
+    /**
+     * Enregistrer un utilisateur,
+     * on passe par la méthode createUser de UserService
+     * @param user
+     * @return
+     */
     @PostMapping(consumes = "application/json", produces = "application/json")
     public User save(@RequestBody User user) {
         return this.userService.createUser(user);
@@ -43,14 +53,29 @@ public class UserController {
 
     }
 
+    /**
+     * Mise à jour de l'utilisateur
+     *
+     * @param id
+     * @param user
+     * @return
+     * @throws Exception
+     */
     @PutMapping("/{id}")
     public User updateUser(@PathVariable Integer id, @RequestBody User user) throws Exception {
         return this.userService.updateUser(id, user);
     }
 
+    /**
+     * Suppression de l'utilisateur
+     * @param id
+     */
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Integer id) {
-        this.userService.deleteById(id);
+        if (userRepository.existsById(id)) {
+            this.userService.deleteById(id);
+
+        }else throw new RuntimeException("User not found");
     }
